requirements.txt
torch
torchvision
hydra-core
torchmetrics
python-dotenv
wandb
torchmetrics
matplotlib
seaborn
scikit-learn